+++ pytest_html/plugin.py
@@ -158,7 +158,8 @@ class HTMLReport(object):
 
             relative_path = '{0}/{1}'.format('assets', asset_file_name)
 
-            with open(asset_path, mode) as f:
+            kwargs = {'encoding': 'utf-8'} if 'b' not in mode else {}
+            with open(asset_path, mode, **kwargs) as f:
                 f.write(content)
             return relative_path